@@ -30,52 +36,72 @@
     "build": "./scripts/build",
     "dev": "./scripts/dev",
     "lint": "./scripts/lint",
+    "make-assets": "./scripts/make-assets",
+    "prettier": "prettier 'src/**/*.[tj]s'",
+    "prettier:fix": "yarn prettier --write",
+    "prettier:check": "yarn prettier --list-different",
     "test": "./scripts/test",
     "prepack": "./scripts/build"
   },
   "dependencies": {
+    "@types/graphql": "^14.0.3",
+    "@types/jsonwebtoken": "<7.2.1",
+    "@types/koa": "^2.0.44",
+    "@types/pg": "^7.4.10",
     "body-parser": "^1.15.2",
-    "chalk": "1.1.3",
-    "commander": "2.9.0",
+    "chalk": "^1.1.3",
+    "commander": "^2.19.0",
     "debug": "^2.3.3",
     "finalhandler": "^1.0.6",
-    "graphql": ">=0.6 <0.13",
+    "graphql": "^0.6.0 || ^0.7.0 || ^0.8.0-b || ^0.9.0 || ^0.10.0 || ^0.11.0 || ^0.12.0 || ^0.13.0 || ^14.0.2",
     "http-errors": "^1.5.1",
     "jsonwebtoken": "^8.0.0",
-    "lodash": ">=3.5 <5",
-    "parseurl": "^1.3.1",
+    "lru-cache": ">=4 <5",
+    "parseurl": "^1.3.2",
     "pg": ">=6.1.0 <8",
     "pg-connection-string": "^0.1.3",
-    "pg-sql2": "2.0.0",
-    "postgraphile-core": "4.0.0-beta.7",
-    "send": "^0.16.1",
+    "pg-sql2": "^2.2.1",
+    "postgraphile-core": "4.3.1",
     "tslib": "^1.5.0"
   },
   "devDependencies": {
+    "@babel/core": "7.0.0",
+    "@babel/plugin-proposal-class-properties": "7.0.0",
+    "@babel/plugin-transform-runtime": "7.0.0",
+    "@babel/preset-env": "7.0.0",
+    "@babel/preset-react": "7.0.0",
+    "@babel/runtime": "7.0.0",
     "@types/chalk": "^0.4.30",
-    "@types/change-case": "2.3.1",
     "@types/commander": ">=2.3.30 <2.12.0",
-    "@types/debug": "0.0.29",
-    "@types/graphql": "^0.8.2",
+    "@types/debug": "^0.0.29",
+    "@types/finalhandler": "^0.0.32",
+    "@types/http-errors": "^1.6.1",
     "@types/jest": "^16.0.4",
-    "@types/jsonwebtoken": "<7.2.1",
-    "@types/lodash": ">=4.14.30 <4.14.57",
-    "@types/node": "^7.0.4",
-    "codemirror-graphql": "^0.6.12",
+    "@types/lru-cache": "^4.1.1",
+    "@types/node": "^10.9.4",
+    "@types/parseurl": "^1.3.1",
+    "babel-loader": "8.0.2",
     "connect": "^3.5.0",
+    "css-loader": "1.0.0",
     "express": "^4.14.0",
-    "graphiql": "^0.11.2",
+    "fastify": "^1.13.0",
+    "html-webpack-inline-source-plugin": "^0.0.10",
+    "html-webpack-plugin": "^3.2.0",
     "jest": "^18.1.0",
+    "koa": "^2.5.1",
+    "koa-compress": "^3.0.0",
+    "koa-mount": "^4.0.0",
     "nodemon": "^1.11.0",
     "pg-minify": "~0.5.3",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-scripts": "1.1.1",
+    "prettier": "^1.15.2",
     "source-map-support": "^0.4.6",
+    "style-loader": "^0.23.0",
     "supertest": "^2.0.1",
     "ts-node": "^2.0.0",
-    "tslint": "^4.2.0",
-    "typescript": "2.1.5"
+    "tslint": "^5.10.0",
+    "tslint-config-prettier": "^1.14.0",
+    "typescript": "^2.9.2",
+    "webpack": "^4.17.2"
   },
   "jest": {
     "transform": {
