@@ -1,8 +1,24 @@
-/* eslint-disable */// Because we use tslint
-import { postgraphile, createPostGraphileSchema, watchPostGraphileSchema, withPostGraphileContext } from './postgraphile'
-import { makePluginHook, PostGraphilePlugin } from './postgraphile/pluginHook'
+import {
+  postgraphile,
+  createPostGraphileSchema,
+  watchPostGraphileSchema,
+  withPostGraphileContext,
+} from './postgraphile';
+import { makePluginHook, PostGraphilePlugin } from './postgraphile/pluginHook';
 
-export default postgraphile
+export {
+  Plugin,
+  Build,
+  Context,
+  SchemaBuilder,
+  SchemaListener,
+  Inflection,
+  Options,
+  postGraphileBaseOverrides,
+  postGraphileClassicIdsOverrides,
+  PostGraphileInflectionPlugin,
+  PostGraphileClassicIdsInflectionPlugin,
+} from 'postgraphile-core';
 
 export {
   postgraphile,
@@ -11,10 +27,13 @@ export {
   withPostGraphileContext,
   makePluginHook,
   PostGraphilePlugin,
-
   // Backwards compatability
   postgraphile as postgraphql,
   createPostGraphileSchema as createPostGraphQLSchema,
   watchPostGraphileSchema as watchPostGraphQLSchema,
   withPostGraphileContext as withPostGraphQLContext,
-}
+};
+
+export * from './interfaces';
+
+export default postgraphile;
