@@ -1,11 +1,11 @@
-const fs = require.requireActual('fs')
+const fs = require.requireActual('fs');
 
 // Mock the writeFile for the export tests
 const writeFile = jest.fn((path, contents, callback) => {
-  callback()
-})
+  callback();
+});
 
 module.exports = {
   ...fs,
   writeFile,
-}
+};
