@@ -1,4 +1,4 @@
-const { GraphQLSchema, GraphQLObjectType, GraphQLInt } = require('graphql')
+const { GraphQLSchema, GraphQLObjectType, GraphQLInt } = require('graphql');
 
 const dummySchema = new GraphQLSchema({
   query: new GraphQLObjectType({
@@ -9,9 +9,9 @@ const dummySchema = new GraphQLSchema({
       },
     },
   }),
-})
+});
 
 module.exports = {
   createPostGraphileSchema: jest.fn(async (a, b, c) => dummySchema),
   watchPostGraphileSchema: jest.fn(async (a, b, c, cb) => cb(dummySchema)),
-}
+};
