@@ -1,10 +1,5 @@
-import postgraphile from './postgraphile'
-import { createPostGraphileSchema, watchPostGraphileSchema } from 'postgraphile-core'
-import withPostGraphileContext from './withPostGraphileContext'
+import postgraphile from './postgraphile';
+import { createPostGraphileSchema, watchPostGraphileSchema } from 'postgraphile-core';
+import withPostGraphileContext from './withPostGraphileContext';
 
-export {
-  postgraphile,
-  createPostGraphileSchema,
-  watchPostGraphileSchema,
-  withPostGraphileContext,
-}
+export { postgraphile, createPostGraphileSchema, watchPostGraphileSchema, withPostGraphileContext };
