@@ -1,17 +1,13 @@
-import { writeFile } from 'fs'
-import { graphql, GraphQLSchema } from 'graphql'
-import { introspectionQuery, printSchema } from 'graphql/utilities'
+import { writeFile } from 'fs';
+import { graphql, GraphQLSchema, introspectionQuery, printSchema } from 'graphql';
 
-async function writeFileAsync(
-  path: string,
-  contents: string,
-): Promise<void> {
+async function writeFileAsync(path: string, contents: string): Promise<void> {
   await new Promise((resolve, reject) => {
     writeFile(path, contents, error => {
-      if (error) reject(error)
-      else resolve()
-    })
-  })
+      if (error) reject(error);
+      else resolve();
+    });
+  });
 }
 
 /**
@@ -20,18 +16,18 @@ async function writeFileAsync(
 export default async function exportPostGraphileSchema(
   schema: GraphQLSchema,
   options: {
-    exportJsonSchemaPath?: string,
-    exportGqlSchemaPath?: string,
+    exportJsonSchemaPath?: string;
+    exportGqlSchemaPath?: string;
   } = {},
 ): Promise<void> {
   // JSON version
   if (typeof options.exportJsonSchemaPath === 'string') {
-    const result = await graphql(schema, introspectionQuery)
-    await writeFileAsync(options.exportJsonSchemaPath, JSON.stringify(result, null, 2))
+    const result = await graphql(schema, introspectionQuery);
+    await writeFileAsync(options.exportJsonSchemaPath, JSON.stringify(result, null, 2));
   }
 
   // Schema language version
   if (typeof options.exportGqlSchemaPath === 'string') {
-    await writeFileAsync(options.exportGqlSchemaPath, printSchema(schema))
+    await writeFileAsync(options.exportGqlSchemaPath, printSchema(schema));
   }
 }
